Tie QueryResult lifetime to that of the Database (#5009)
It should be safe to use QueryResults after the connection has been destroyed, but it isn't after the Database has been destroyed
Fixed some warnings and also bumped the MSRV to 1.67 since clippy
pointed out that checked_ilog10 has only been stable since then

+++ tools/rust_api/src/database.rs
@@ -51,7 +51,7 @@ impl Default for SystemConfig {
             // This is a little weird, but it's a temporary interface
             max_db_size: u32::MAX as u64,
             auto_checkpoint: true,
-            checkpoint_threshold: -1 as i64,
+            checkpoint_threshold: -1_i64,
         }
     }
 }
